+++ dflib/src/main/java/org/dflib/exp/agg/DecimalAggregators.java
@@ -6,13 +6,21 @@ import org.dflib.agg.Percentiles;
 import org.dflib.builder.ObjectAccum;
 
 import java.math.BigDecimal;
+import java.math.MathContext;
+import java.math.RoundingMode;
 
-import static org.dflib.Exp.*;
+import static org.dflib.Exp.$decimal;
+import static org.dflib.Exp.count;
 
 
 public class DecimalAggregators {
 
     private static final DecimalExp avg = $decimal(0).sum().div(count());
+
+    // TODO: unify with the context in the DecimalExpFactory
+    static MathContext op1Context(BigDecimal n) {
+        return new MathContext(Math.max(15, 1 + n.scale()), RoundingMode.HALF_UP);
+    }
 
     public static Series<BigDecimal> cumSum(Series<BigDecimal> s) {
 
@@ -78,5 +86,40 @@ public class DecimalAggregators {
      */
     public static BigDecimal avg(Series<BigDecimal> s) {
         return s.size() == 0 ? BigDecimal.ZERO : avg.reduce(s);
+    }
+
+    /**
+     * @since 1.3.0
+     */
+    public static BigDecimal variance(Series<BigDecimal> s, boolean usePopulationVariance) {
+
+        int size = s.size();
+        BigDecimal avg = DecimalAggregators.avg(s);
+        MathContext mc = op1Context(avg);
+        BigDecimal denominator = new BigDecimal(usePopulationVariance ? size : size - 1);
+
+        // TODO: ignoring a possibility of nulls... e.g., numpy throws when calculating a variance of array with Nones
+        //  Should we be smarter?
+
+        BigDecimal acc = BigDecimal.ZERO;
+        for (int i = 0; i < size; i++) {
+            BigDecimal x = s.get(i);
+
+
+            BigDecimal dev = x.subtract(avg, mc);
+            BigDecimal square = dev.multiply(dev, mc);
+
+            acc = acc.add(square, mc);
+        }
+
+        return acc.divide(denominator, mc);
+    }
+
+    /**
+     * @since 1.3.0
+     */
+    public static BigDecimal stdDev(Series<BigDecimal> s, boolean usePopulationStdDev) {
+        BigDecimal variance = variance(s, usePopulationStdDev);
+        return variance.sqrt(op1Context(variance));
     }
 }