+++ dflib/src/main/java/org/dflib/exp/num/DoubleExpFactory.java
@@ -130,6 +130,16 @@ public class DoubleExpFactory extends NumericExpFactory {
     }
 
     @Override
+    public NumExp<?> variance(Exp<? extends Number> exp, boolean usePopulationVariance) {
+        return new DoubleReduceExp1<>("variance", exp, s -> DoubleAggregators.variance(s, usePopulationVariance), null);
+    }
+
+    @Override
+    public NumExp<?> stdDev(Exp<? extends Number> exp, boolean usePopulationStdDev) {
+        return new DoubleReduceExp1<>("stdDev", exp, s -> DoubleAggregators.stdDev(s, usePopulationStdDev), null);
+    }
+
+    @Override
     public NumExp<Long> round(Exp<? extends Number> exp) {
         return LongExp1.mapVal("round", cast(exp), n -> Math.round(n));
     }