+++ dflib/src/test/java/org/dflib/IntSeries_StdDevTest.java
@@ -11,18 +11,12 @@ public class IntSeries_StdDevTest {
 
         IntSeries series = Series.ofInt(3, 28, 15, -4, 3, 11);
 
-        /**
-         * Validated results with python:
-         *
-         * In [1]: import numpy as np
-         *
-         * In [2]: a = np.array([3, 28, 15, -4, 3, 11])
-         *
-         * In [3]: np.std(a)
-         *
-         * Out[3]: 10.338708279513881
-         *
-         */
+        // Validated results with Python:
+        // In [1]: import numpy as np
+        // In [2]: a = np.array([3, 28, 15, -4, 3, 11])
+        // In [3]: np.std(a)
+        // Out[3]: 10.338708279513881
+
         assertEquals(10.338708279513881d, series.stdDev(true));
         assertEquals(10.338708279513881d, series.stdDev());
     }
@@ -30,19 +24,14 @@ public class IntSeries_StdDevTest {
     @Test
     public void stdDev_Sample() {
 
-        final IntSeries series = Series.ofInt(3, 28, 15, -4, 3, 11);
+        IntSeries series = Series.ofInt(3, 28, 15, -4, 3, 11);
 
-        /**
-         * Validated results with python:
-         *
-         * In [1]: import numpy as np
-         *
-         * In [2]: a = np.array([3, 28, 15, -4, 3, 11])
-         *
-         * In [3]: np.std(a, ddof=1)
-         *
-         * Out[3]: 11.325487480310358
-         */
+        // Validated results with Python:
+        // In [1]: import numpy as np
+        // In [2]: a = np.array([3, 28, 15, -4, 3, 11])
+        // In [3]: np.std(a, ddof=1)
+        // Out[3]: 10.338708279513881
+
         double stdDev = series.stdDev(false);
         assertEquals(11.325487480310358d, stdDev);
     }